Move SignInModal to shared
